requirements.txt
flask
feedparser
requests
beautifulsoup4
anthropic
python-dotenv
